==> benchmark/0.create-experiment-metadata.ipynb <==
import pandas as pd
import os
import glob
import numpy as np

col_order = [
    "Batch",
    "Plate_Map_Name",
    "Assay_Plate_Barcode",
    "Perturbation",
    "Cell_type",
    "Time",
    "Density",
    "Antibiotics",
    "Cell_line",
    "Time_delay",
    "Times_imaged",
    "Anomaly",
    "Number_of_images",
]

antibiotics_df = pd.DataFrame(
    {
        "Assay_Plate_Barcode": ["BR00118049", "BR00117002", "BR00117001"],
        "Antibiotics": ["Blasticidin", "Puromycin", "Puromycin"],
    }
)

density_df = pd.DataFrame(
    {
        "Assay_Plate_Barcode": ["BR00117054", "BR00117055", "BR00117008", "BR00117009"],
        "Density": ["120", "120", "80", "80"],
    }
)

cell_line_df = pd.DataFrame(
    {
        "Assay_Plate_Barcode": ["BR00117052", "BR00117053", "BR00117050", "BR00117051"],
        "Cell_line": ["Cas9", "Cas9", "Cas9", "Cas9"],
    }
)

anomalous_plates = {
    "Phalloidin": [
        "BR00118050",
        "BR00117006",
        "BR00118049",
        "BR00118039",
        "BR00118040",
        "BR00118041",
        "BR00118042",
        "BR00118043",
        "BR00118044",
        "BR00118045",
        "BR00118046",
        "BR00118047",
        "BR00118048",
    ],
    "WGA": [
        "BR00116997",
        "BR00116998",
        "BR00116999",
        "BR00116996",
        "BR00116991",
        "BR00116992",
        "BR00116993",
        "BR00116994",
        "BR00116995",
        "BR00117024",
        "BR00117025",
        "BR00117026",
    ],
    "Mitotracker": ["BR00117013", "BR00117010"],
}

# Read barcode_platemap files
path = os.path.join("../metadata/platemaps/*/barcode_platemap.csv")
files = glob.glob(path)
df = pd.concat(
    (
        (
            pd.read_csv(_)
            .assign(Batch=_.split("/")[3])
            .assign(Anomaly="none")
            .assign(Time_delay="Day0")
            .assign(Times_imaged="1")
            .assign(Time="0")
        )
        for _ in files
    ),
    ignore_index=True,
)

# Read number of images csv
path = os.path.join("input/2020*.csv")
files = glob.glob(path)
images_df = pd.concat((pd.read_csv(_) for _ in files), ignore_index=True)

df["Perturbation"] = np.asarray(list(df.Plate_Map_Name.str.split("_").values))[:, 1]

# Add antibiotics column
df = df.merge(antibiotics_df, on="Assay_Plate_Barcode", how="left")
df.Antibiotics.replace(np.nan, "absent", inplace=True)
df.query('Antibiotics!="absent"').head()

# Add density column
df = df.merge(density_df, on="Assay_Plate_Barcode", how="left")
df.Density.replace(np.nan, "100", inplace=True)
df.query('Density!="100"').head()

# Add Cell line column
df = df.merge(cell_line_df, on="Assay_Plate_Barcode", how="left")
df.loc[df.Perturbation == "crispr", "Cell_line"] = "Cas9"
df.Cell_line.replace(np.nan, "Parental", inplace=True)
df.query('Cell_line=="Cas9"').head(50)

# Fill the Time delay column
df.loc[df.Batch.str.contains("Day1"), "Time_delay"] = "Day1"
df.loc[df.Batch.str.contains("Day4"), "Time_delay"] = "Day4"
df.loc[df.Batch.str.contains("2Weeks"), "Time_delay"] = "Week2"
df.loc[df.Batch.str.contains("4Weeks"), "Time_delay"] = "Week4"
df.query('Time_delay!="Day0"').head(50)

# Fill the Times imaged column
df.loc[df.Assay_Plate_Barcode.str.endswith("A"), "Times_imaged"] = "2"
df.loc[df.Assay_Plate_Barcode.str.endswith("B"), "Times_imaged"] = "3"
df.loc[df.Assay_Plate_Barcode.str.endswith("C"), "Times_imaged"] = "4"
df.loc[df.Assay_Plate_Barcode.str.endswith("D"), "Times_imaged"] = "5"
df.loc[df.Assay_Plate_Barcode.str.endswith("E"), "Times_imaged"] = "6"
df.loc[df.Assay_Plate_Barcode.str.endswith("F"), "Times_imaged"] = "7"

df.loc[df.Time_delay == "Week2", "Times_imaged"] = "2"
df.loc[df.Time_delay == "Week4", "Times_imaged"] = "3"

df.query('Times_imaged=="3"').head(50)

# Fill anomalies column
for anomaly in anomalous_plates:
    for plate in anomalous_plates[anomaly]:
        df.loc[df.Assay_Plate_Barcode == plate, "Anomaly"] = anomaly
df.query('Anomaly!="none"').head(50)

# Add Number of images column
df = df.merge(images_df, on=["Batch", "Assay_Plate_Barcode"], how="left")
df.head()

# Create Cell type column
df["Cell_type"] = list(df.Description.str[0:4])
df.head()

# Create Time column
df.loc[df.Description.str.contains("24"), "Time"] = "24"
df.loc[df.Description.str.contains("48"), "Time"] = "48"
df.loc[df.Description.str.contains("96"), "Time"] = "96"
df.loc[df.Description.str.contains("144"), "Time"] = "144"
df.loc[df.Time == "0"]

# Sort rows and columns
df = df.sort_values(
    by=["Batch", "Plate_Map_Name", "Assay_Plate_Barcode"], ignore_index=True
)[col_order]

df.to_csv("output/experiment-metadata.tsv", index=False, sep="\t")
