==> benchmark/1.3.rename-output-columns.ipynb <==
import pandas as pd

Rename column names in the mAP files

q value in column names will be converted to corrected p value.

def rename_columns(df):
    df.rename(
        columns={
            "q_value": "corrected_p_value",
            "nlog10qvalue": "nlog10correctedpvalue",
            "above_q_threshold": "above_corrected_p_threshold",
        },
        inplace=True,
    )
    return df

filenames = [
    "cellprofiler_replicability_map.csv",
    "cellprofiler_matching_map.csv",
    "cellprofiler_gene_compound_matching_map.csv",
    "cellprofiler_replicability_orf_same_map.csv",
    "cellprofiler_replicability_orf_different_map.csv"
]

for filename in filenames:
    df = pd.read_csv(f"output/{filename}")
    df = rename_columns(df)
    df.to_csv(f"output/{filename}", index=False)
