==> benchmark/7.create-tables.ipynb <==
import pandas as pd

Replicability

replicability_pvalue = pd.read_csv("output/replicability_pvalue.csv")

replicability_fr = pd.read_csv("output/cellprofiler_replicability_fr.csv")

print(
    replicability_pvalue.merge(
        replicability_fr, on=["Modality", "Cell", "time"]
    )
    .sort_values(by=["Modality", "Cell", "time"])
    .to_markdown(index=False)
)

Same perturbation matching

matching_pvalue = pd.read_csv("output/matching_pvalue.csv")

matching_fr = pd.read_csv("output/cellprofiler_matching_fr.csv")

print(
    matching_pvalue.merge(
        matching_fr, on=["Modality", "Cell", "time"]
    )
    .sort_values(by=["Modality", "Cell", "time"])
    .to_markdown(index=False)
)

Different perturbation matching

gene_compound_matching_pvalue = pd.read_csv("output/gene_compound_matching_pvalue.csv")

gene_compound_matching_fr = pd.read_csv("output/cellprofiler_gene_compound_matching_fr.csv")

print(
    gene_compound_matching_pvalue.merge(
        gene_compound_matching_fr,
        on=["Modality1", "Modality2", "Cell"],
    )
    .sort_values(by=["Modality1", "Modality2", "Cell"])
    .to_markdown(index=False)
)

Different pertubation matching CRISPR vs. ORF

gene_compound_matching_crispr_orf_pvalue = pd.read_csv("output/crispr_orf_pvalue.csv")

print(
    gene_compound_matching_crispr_orf_pvalue.sort_values(
        by=["compound-crispr", "compound-orf", "Cell"]
    )[
        [
            "t-test",
            "Cell",
            "compound-crispr",
            "compound-orf",
            "statistic",
            "pvalue",
        ]
    ].sort_values(by=["Cell","compound-crispr", "compound-orf"]).to_markdown(
        index=False
    )
)

Compound-genetic perturbation cosine similarity lists

cols = [
    "Cell",
    "Genetic_Perturbation",
    "pert_iname",
    "Metadata_matching_target",
    "moa_list",
    "cosine_sim",
]

cosine_sim = pd.read_csv("output/compound_genetic_perturbation_cosine_similarity.csv")
metadata = (
    pd.read_csv(
        "../metadata/external_metadata/JUMP-Target-1_compound_metadata.tsv",
        sep="\t",
        usecols=["broad_sample", "pert_iname"],
    )
    .merge(
        pd.read_csv(
            "input/JUMP-Target-1_compound_metadata_additional_annotations.tsv",
            sep="\t",
            usecols=["broad_sample", "moa_list"],
        )
    )
    .rename(columns={"broad_sample": "Metadata_broad_sample"})
)

cosine_sim = cosine_sim.merge(metadata, on="Metadata_broad_sample")

for cell_type in cosine_sim.Cell.unique():
    filtered_df = cosine_sim.query("Cell==@cell_type")
    top_df = filtered_df.nlargest(10, "cosine_sim")
    bottom_df = filtered_df.nsmallest(10, "cosine_sim").sort_values(
        by="cosine_sim", ascending=False
    )
    df = pd.concat([top_df, bottom_df], join="inner")
    print(df[cols].to_markdown(index=False))
