==> benchmark/9.count-datapoints.ipynb <==
import pandas as pd

Read and print the mAP files

replicability_df = (
    pd.read_csv('output/cellprofiler_replicability_map.csv')
    .assign(Description= lambda x: x["Modality"] + "_" + x["time"] + "_" + x["Cell"])
    .assign(Type="Replicate retrieval")
)
replicability_df = (
    replicability_df.groupby(["Type", "Description"])
    .mean_average_precision.count()
    .reset_index()
    .rename(columns={"mean_average_precision": "n_data_points"})
)

matching_df = (
    pd.read_csv('output/cellprofiler_matching_map.csv')
    .assign(Description= lambda x: x["Modality"] + "_" + x["time"] + "_" + x["Cell"])
    .assign(Type="Sister perturbation retrieval")
)
matching_df = (
    matching_df.groupby(["Type", "Description"])
    .mean_average_precision.count()
    .reset_index()
    .rename(columns={"mean_average_precision": "n_data_points"})
)

gene_compound_matching_df = (
    pd.read_csv('output/cellprofiler_gene_compound_matching_map.csv')
    .assign(Description= lambda x: x["Modality1"] + "_" + x["Modality2"] + "_" + x["Cell"])
    .assign(Type="Gene compound retrieval")
)
gene_compound_matching_df = (
    gene_compound_matching_df.groupby(["Type", "Description"])
    .mean_average_precision.count()
    .reset_index()
    .rename(columns={"mean_average_precision": "n_data_points"})
)

Combine dataframes

print(
    pd.concat(
        [
            replicability_df,
            matching_df,
            gene_compound_matching_df
        ]
    ).to_markdown(
        index=False
    )
)
