==> benchmark/10.analyze-brd4.ipynb <==
import pandas as pd

matching_df = pd.read_csv("output/cellprofiler_matching_map.csv")

print("Do compounds that target BRD4 look similar?")
matching_df.query('Metadata_matching_target=="BRD4" and Modality=="compound"')[
    ["Modality", "Cell", "time", "Metadata_matching_target", "above_corrected_p_threshold"]
].reset_index(drop=True).sort_values(by=["Modality", "Cell", "time"])

cosine_sim = pd.read_csv("output/compound_genetic_perturbation_cosine_similarity.csv")

metadata = (
    pd.read_csv(
        "../metadata/external_metadata/JUMP-Target-1_compound_metadata.tsv",
        sep="\t",
        usecols=["broad_sample", "pert_iname"],
    )
    .merge(
        pd.read_csv(
            "input/JUMP-Target-1_compound_metadata_additional_annotations.tsv",
            sep="\t",
            usecols=["broad_sample", "moa_list"],
        )
    )
    .rename(columns={"broad_sample": "Metadata_broad_sample"})
)

cosine_sim_brd4 = cosine_sim.query('Metadata_matching_target=="BRD4"').merge(
    metadata, on="Metadata_broad_sample"
)
cosine_sim_bi2536 = cosine_sim.merge(metadata, on="Metadata_broad_sample").query(
    'pert_iname=="BI-2536"'
)

print("Cosine similarity of BI-2536 to its gene targets")
cosine_sim_bi2536.sort_values(by="cosine_sim", ascending=False).groupby(
    ["Metadata_matching_target", "Genetic_Perturbation", "Cell"]
).cosine_sim.median().reset_index()

print("Cosine similarity of BRD4 to the compounds that target it")
cosine_sim_brd4.sort_values(by="cosine_sim", ascending=False).groupby(
    ["pert_iname", "Genetic_Perturbation", "Cell"]
).cosine_sim.median().reset_index()

Check if PFI-1 and TG-101348 are replicable

other_brd4 = metadata.query('pert_iname=="PFI-1" or pert_iname=="TG-101348"')

replicate_retrieval_other_brd4 = pd.read_csv("output/cellprofiler_replicability_map.csv").merge(
    other_brd4, on="Metadata_broad_sample"
)
replicate_retrieval_other_brd4

all_brd4 = metadata.query('pert_iname=="PFI-1" or pert_iname=="TG-101348" or pert_iname=="BI-2536"')

replicate_retrieval_all_brd4 = pd.read_csv("output/cellprofiler_replicability_map.csv").merge(
    all_brd4, on="Metadata_broad_sample"
)
replicate_retrieval_all_brd4
